+++ backend/schemas/inventory_item_audit.py
@@ -1,13 +1,14 @@
 from pydantic import BaseModel
 from datetime import datetime
 from typing import Optional
+from decimal import Decimal
 
 class InventoryItemAuditBase(BaseModel):
     inventory_item_id: int
     change_type: str
-    change_amount: float
-    old_quantity: float
-    new_quantity: float
+    change_amount: Decimal
+    old_quantity: Decimal
+    new_quantity: Decimal
     changed_by: Optional[str] = None
     note: Optional[str] = None
     tenant_id: Optional[str] = None
